mem: fix -Werror=format build failure in hg_arena_init()
HG_CELL_HDR is built from ROUNDTO (sizeof-derived), so it has type
size_t, not int. Printing it with %d tripped -Wformat and, since the
build runs with -Werror, aborted the whole compile at mem/hg_arena.o.

Use %zu, matching how f_malloc.c, q_malloc.c and hp_malloc.c already
print their sizeof-derived values. %zu is also correct on 32-bit, where
ROUNDTO is sizeof(void *).

+++ mem/hg_arena.c
@@ -235,9 +235,9 @@ int hg_arena_init(struct hg_block *hb, unsigned long hdr_size)
 		hb->size2class[idx] = (unsigned char)c;   /* HG_NCLASSES = oversize */
 	}
 
-	LM_DBG("%s arena ready: %d classes, %u B to %u B cells (header=%lu B)\n",
+	LM_DBG("%s arena ready: %d classes, %u B to %u B cells (header=%zu B)\n",
 		hb->name, HG_NCLASSES, cell_sizes[0], cell_sizes[HG_NCLASSES-1],
-		(unsigned long)HG_CELL_HDR);
+		HG_CELL_HDR);
 	return 0;
 }
